backend/requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.30,<0.31
python-dotenv>=1,<2
openai>=1.37.1,<2
langchain>=0.2.7,<0.3
langchain-openai>=0.1.14,<0.2
pinecone-client>=5,<6
httpx>=0.27,<0.28
pydantic>=2.7.4,<3
